Adapted the classes to calculate the spectra for the vibrational corrections. Removed the use of Float and now one uses always Doubles.

+++ src/main/java/uk/ac/ebi/nmr/execs/NMRDBSimulatorAllConf.java
@@ -1,5 +1,6 @@
 package uk.ac.ebi.nmr.execs;
 
+import org.openscience.cdk.interfaces.IAtom;
 import org.openscience.cdk.interfaces.IAtomContainer;
 import org.openscience.cdk.interfaces.IChemFile;
 import org.openscience.cdk.interfaces.IChemModel;
@@ -8,6 +9,7 @@ import org.openscience.cdk.tools.manipulator.ChemFileManipulator;
 import uk.ac.ebi.nmr.MagneticPropTool;
 import uk.ac.ebi.nmr.SpinSystem;
 import uk.ac.ebi.nmr.io.G09OutputReader;
+import uk.ac.ebi.nmr.io.L777OutputReader;
 import uk.ac.ebi.nmr.io.NMRDBSimulatorWriter;
 
 import java.io.BufferedReader;
@@ -16,6 +18,7 @@ import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.FilenameFilter;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.regex.Pattern;
 
@@ -27,7 +30,6 @@ import java.util.regex.Pattern;
  * To change this template use File | Settings | File Templates.
  */
 public class NMRDBSimulatorAllConf {
-
 
     public static void main(String[] args) throws Exception {
 
@@ -45,7 +47,7 @@ public class NMRDBSimulatorAllConf {
         FilenameFilter gaussianFilter = new FilenameFilter() {
             @Override
             public boolean accept(File file, String filename) {
-                Pattern REGEX_GAUSSIAN_FILE = Pattern.compile("conf\\_\\d+(\\_solvent)?\\.log");
+                Pattern REGEX_GAUSSIAN_FILE = Pattern.compile("conf\\_\\d+(\\_solvent)?(-vibcorr)?\\.log");
                 return REGEX_GAUSSIAN_FILE.matcher(filename).find();
             }
         };
@@ -56,28 +58,64 @@ public class NMRDBSimulatorAllConf {
 
         List<IAtomContainer> vacuumSimulations = new ArrayList<IAtomContainer>();
         List<IAtomContainer> solventSimulations = new ArrayList<IAtomContainer>();
+        List<IAtomContainer> vacuumVibcorrSimulations = new ArrayList<IAtomContainer>();
+        List<IAtomContainer> solventVibcorrSimulations = new ArrayList<IAtomContainer>();
         System.out.println(gaussianFiles.length);
         for (File file : gaussianFiles) {
             BufferedReader bufferR = new BufferedReader(new FileReader(file));
             G09OutputReader g09OutputReader = new G09OutputReader(bufferR);
+            L777OutputReader l777OutputReader = new L777OutputReader(bufferR);
             IChemFile chemFile = new ChemFile();
-            chemFile = g09OutputReader.read(chemFile);
+            if(file.getName().contains("vibcorr")){
+                chemFile = l777OutputReader.read(chemFile);
+            }else{
+                chemFile = g09OutputReader.read(chemFile);
+            }
             // the reader will read a sequence of atomcontainer and only the last one will have all the information
             IChemModel model = ChemFileManipulator.getAllChemModels(chemFile)
                     .get(ChemFileManipulator.getAllChemModels(chemFile).size() - 1);
-            System.out.println((Boolean) model.getProperty(G09OutputReader.NORMAL_TERMINATION));
+
+            if(file.getName().contains("vibcorr")){
+                model.setProperty(G09OutputReader.NORMAL_TERMINATION,
+                        (Boolean) model.getProperty(L777OutputReader.NORMAL_TERMINATION));
+            }
+
+            System.out.println("processing file: "+file.getName());
+            System.out.println("Normal termination: "+
+                    ((Boolean) model.getProperty(G09OutputReader.NORMAL_TERMINATION)));
+
             if ((Boolean) model.getProperty(G09OutputReader.NORMAL_TERMINATION)) {
-                System.out.println("processing file: "+file.getName());
                 IAtomContainer atomContainer = ChemFileManipulator.getAllAtomContainers(chemFile)
                         .get(ChemFileManipulator.getAllAtomContainers(chemFile).size() - 1);
-                System.out.println("Energy of the geometry: " + model.getProperty(G09OutputReader.STRUCTURE_ENERGY));
-                atomContainer.setProperty(G09OutputReader.STRUCTURE_ENERGY,
-                        Double.parseDouble((String) model.getProperty(G09OutputReader.STRUCTURE_ENERGY)));
+                if(file.getName().contains("vibcorr")){
+                    for(IAtom atom : atomContainer.atoms()){
+                        atom.setProperty(G09OutputReader.MAGNETIC_TENSOR,
+                                (Double) atom.getProperty(L777OutputReader.MAGNETIC_TENSOR));
+                        atom.setProperty(G09OutputReader.COUPLING_CONSTANTS,
+                                (HashMap<IAtom,Double>) atom.getProperty(L777OutputReader.COUPLING_CONSTANTS));
+                    }
+                    System.out.println(model.getProperty(L777OutputReader.STRUCTURE_ENERGY));
+                    atomContainer.setProperty(G09OutputReader.STRUCTURE_ENERGY,
+                            Double.parseDouble((String) model.getProperty(L777OutputReader.STRUCTURE_ENERGY)));
+                } else {
+                    atomContainer.setProperty(G09OutputReader.STRUCTURE_ENERGY,
+                            Double.parseDouble((String) model.getProperty(G09OutputReader.STRUCTURE_ENERGY)));
+                }
+                System.out.println("Energy of the geometry: " +
+                        atomContainer.getProperty(G09OutputReader.STRUCTURE_ENERGY));
                 //sort the atomcontainer
                 if (file.getName().contains("solvent")) {
-                    solventSimulations.add(atomContainer);
+                    if(file.getName().contains("vibcorr")){
+                        solventVibcorrSimulations.add(atomContainer);
+                    } else {
+                        solventSimulations.add(atomContainer);
+                    }
                 } else {
-                    vacuumSimulations.add(atomContainer);
+                    if(file.getName().contains("vibcorr")){
+                        vacuumVibcorrSimulations.add(atomContainer);
+                    } else {
+                        vacuumSimulations.add(atomContainer);
+                    }
                 }
             }
             // write down the NMR data
@@ -86,16 +124,33 @@ public class NMRDBSimulatorAllConf {
             throw new Exception("No simulations in vacuum conditions found");
         if(solventSimulations.size()==0)
             throw new Exception("No simulations in solvent conditions found");
+        if(solventVibcorrSimulations.size()==0)
+            throw new Exception("No simulations in solvent and vibrational corrections conditions found");
+        if(vacuumVibcorrSimulations.size()==0)
+            throw new Exception("No simulations in vacuum and vibrational corrections conditions found");
 
         SpinSystem averageSpinSystemVacuum = new MagneticPropTool().calculateAverageHSpinSystem(vacuumSimulations);
         SpinSystem averageSpinSystemSolvent = new MagneticPropTool().calculateAverageHSpinSystem(solventSimulations);
+        SpinSystem averageSpinSystemVacuumVibbcorr = new MagneticPropTool()
+                .calculateAverageHSpinSystem(vacuumVibcorrSimulations);
+        SpinSystem averageSpinSystemSolventVibbcorr = new MagneticPropTool()
+                .calculateAverageHSpinSystem(solventVibcorrSimulations);
 
         NMRDBSimulatorWriter nmrdbSimulatorWriter = new NMRDBSimulatorWriter(new FileWriter(new File(
-                workingDir+"/averageSpectra-vacuum-simulator.txt")));
+                workingDir+"/avgSpectra-vacuum-sim.txt")));
         nmrdbSimulatorWriter.write(averageSpinSystemVacuum);
 
         nmrdbSimulatorWriter = new NMRDBSimulatorWriter(new FileWriter(new File(
-                workingDir+"/averageSpectra-solvent-simulator.txt")));
+                workingDir+"/avgSpectra-vacuum-vibcorr-sim.txt")));
+        nmrdbSimulatorWriter.write(averageSpinSystemVacuumVibbcorr);
+
+        nmrdbSimulatorWriter = new NMRDBSimulatorWriter(new FileWriter(new File(
+                workingDir+"/avgSpectra-solvent-sim.txt")));
         nmrdbSimulatorWriter.write(averageSpinSystemSolvent);
+
+        nmrdbSimulatorWriter = new NMRDBSimulatorWriter(new FileWriter(new File(
+                workingDir+"/avgSpectra-solvent-vibcorr-sim.txt")));
+        nmrdbSimulatorWriter.write(averageSpinSystemSolventVibbcorr);
+
     }
 }
